(RE-4200) Add osx 10.9 platform config
This has a platform config for osx 10.9. OSX is special in that it
doesn't have pkgconfig available, so the provisioning includes building
the pkgconfig package from source. We also lay down the libxml2
pkgconfig file which is conspicuously absent from the default osx
install.

Servicetype and service dir are set because the build will fail without
them, but they are almost certainly wrong.